# smartrose-api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SMART-ROSE API",
    "version": "1.0.0",
    "description": "API Sistem Pelaporan Insiden Keselamatan Pasien untuk Puskesmas"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base API path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/reports": {
      "post": {
        "summary": "Buat laporan insiden baru (publik, tanpa login)",
        "description": "Endpoint publik untuk pelaporan insiden tanpa perlu akun. Laporan otomatis berstatus SUBMITTED dan bersifat anonim. Wajib menyertakan captchaToken dan captchaJawaban yang diperoleh dari GET /captcha.\n",
        "tags": [
          "Reports"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jenisInsiden",
                  "tanggalKejadian",
                  "lokasi",
                  "unitKerja",
                  "kronologi",
                  "dampak",
                  "gradingAwal",
                  "captchaToken",
                  "captchaJawaban"
                ],
                "properties": {
                  "jenisInsiden": {
                    "type": "string",
                    "enum": [
                      "KTD",
                      "KNC",
                      "KTC",
                      "KPC",
                      "SENTINEL"
                    ]
                  },
                  "tanggalKejadian": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "lokasi": {
                    "type": "string"
                  },
                  "unitKerja": {
                    "type": "string"
                  },
                  "kronologi": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 5000
                  },
                  "dampak": {
                    "type": "string"
                  },
                  "gradingAwal": {
                    "type": "string",
                    "enum": [
                      "HIJAU",
                      "BIRU",
                      "KUNING",
                      "MERAH"
                    ]
                  },
                  "captchaToken": {
                    "type": "string",
                    "description": "Token UUID dari GET /captcha"
                  },
                  "captchaJawaban": {
                    "type": "string",
                    "description": "Jawaban soal matematika dari GET /captcha"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Laporan berhasil dibuat, mengembalikan trackingNumber"
          },
          "400": {
            "description": "Validasi gagal atau captcha salah"
          },
          "429": {
            "description": "Terlalu banyak permintaan"
          }
        }
      }
    },
...
    "/admin/users/pending": {
      "get": {
        "summary": "Get all pending users",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending users"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/users": {
      "post": {
        "summary": "Create a new admin user (ADMIN_UTAMA only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nama",
                  "email",
                  "noPegawai",
                  "unitKerja"
                ],
                "properties": {
                  "nama": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "noPegawai": {
                    "type": "string"
                  },
                  "unitKerja": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin user created successfully (returns plain text password once)"
          },
          "400": {
            "description": "Validation error or email/noPegawai already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Not ADMIN_UTAMA)"
          }
        }
      },
      "get": {
        "summary": "Get list of users (Admin)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN"
              ]
            },
            "description": "Filter by role"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or email"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/admin/users/{id}/approve": {
      "post": {
        "summary": "Approve a pending user",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User approved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/admin/users/{id}/reject": {
      "post": {
        "summary": "Reject a pending user",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User rejected successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/admin/reports/export": {
      "get": {
        "summary": "Export reports list to PDF or Excel",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "excel",
                "pdf"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download file"
          },
          "400": {
            "description": "Format invalid"
          }
        }
      }
    },
    "/admin/reports": {
      "get": {
        "summary": "List all reports for admin",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of reports"
          }
        }
      },
      "post": {
        "summary": "Create report manually",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/admin/reports/{id}": {
      "get": {
        "summary": "Get report detail for admin",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report detail"
          }
        }
      },
      "put": {
        "summary": "Update report as admin",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      }
    },
    "/admin/reports/{id}/assign": {
      "post": {
        "summary": "Assign report",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assigned"
          }
        }
      }
    },
    "/admin/reports/{id}/archive": {
      "post": {
        "summary": "Archive report",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archived"
          }
        }
      }
    },
    "/admin/reports/{id}/hard": {
      "delete": {
        "summary": "Hard delete report",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/admin/reports/{id}/regrade": {
      "put": {
        "summary": "Regrade laporan (ubah grading risiko dan hitung ulang deadline SLA)",
        "tags": [
          "Admin Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "gradingBaru",
                  "alasan"
                ],
                "properties": {
                  "gradingBaru": {
                    "type": "string",
                    "enum": [
                      "HIJAU",
                      "BIRU",
                      "KUNING",
                      "MERAH"
                    ]
                  },
                  "alasan": {
                    "type": "string",
                    "minLength": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grading berhasil diubah, deadline baru dihitung dari sekarang (UTC)"
          },
          "400": {
            "description": "Input tidak valid"
          },
          "404": {
            "description": "Laporan tidak ditemukan"
          }
        }
      }
    },
    "/admin/jobs/sla-check": {
      "post": {
        "summary": "Trigger SLA check secara manual (ADMIN_UTAMA only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "SLA check berhasil dijalankan"
          },
          "403": {
            "description": "Hanya ADMIN_UTAMA yang bisa trigger ini"
          }
        }
      }
    },
...
  },
...
}